requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-mysql==5.6.1
PyMySQL==1.1.1
pyspark==3.5.1
hdfs==2.7.3